[fix] removed unused functions on certificate helper
